+++ .prettierrc.js
@@ -5,4 +5,5 @@ module.exports = {
 	useTabs: true,
 	tabWidth: 2,
 	printWidth: 100,
+	endOfLine: 'lf',
 };